requirements.txt
fastapi==0.117.1
pydantic==2.11.9
pymongo==4.15.1
tavily-python==0.7.12
langchain==0.3.27
langchain-openai==0.3.33
networkx==3.5
matplotlib==3.10.6
python-dotenv==1.1.1
uvicorn==0.37.0